requirements.txt
# Core ML framework
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
protobuf==3.19.6
h5py==3.7.0

# Data handling
pandas==1.5.3
scikit-learn==1.1.3

# Image processing
pillow==9.5.0

# Visualization
matplotlib==3.6.3
seaborn==0.12.2
altair==4.2.2

# Web app
streamlit==1.20.0

# Utility
opencv-python==4.6.0.66
packaging==21.3
